+++ tests/test_entity.py
@@ -28,7 +28,7 @@ async def test_custom_entity_callback(
     central_local_factory: helper.CentralUnitLocalFactory,
 ) -> None:
     """Test CeSwitch."""
-    central, mock_client = await central_local_factory.get_central(TEST_DEVICES)
+    central, mock_client = await central_local_factory.get_default_central(TEST_DEVICES)
     switch: CeSwitch = cast(CeSwitch, await get_custom_entity(central, "VCU2128127", 4))
     assert switch.usage == HmEntityUsage.CE_PRIMARY
 
@@ -70,7 +70,7 @@ async def test_generic_entity_callback(
     central_local_factory: helper.CentralUnitLocalFactory,
 ) -> None:
     """Test CeSwitch."""
-    central, mock_client = await central_local_factory.get_central(TEST_DEVICES)
+    central, mock_client = await central_local_factory.get_default_central(TEST_DEVICES)
     switch: HmSwitch = cast(
         HmSwitch, await get_generic_entity(central, "VCU2128127:4", "STATE")
     )
@@ -114,7 +114,7 @@ async def test_load_custom_entity(
     central_local_factory: helper.CentralUnitLocalFactory,
 ) -> None:
     """Test load custom_entity."""
-    central, mock_client = await central_local_factory.get_central(TEST_DEVICES)
+    central, mock_client = await central_local_factory.get_default_central(TEST_DEVICES)
     switch: HmSwitch = cast(HmSwitch, await get_custom_entity(central, "VCU2128127", 4))
     await switch.load_entity_value(call_source=HmCallSource.MANUAL_OR_SCHEDULED)
     assert mock_client.method_calls[-2] == call.get_value(
@@ -136,7 +136,7 @@ async def test_load_generic_entity(
     central_local_factory: helper.CentralUnitLocalFactory,
 ) -> None:
     """Test load generic_entity."""
-    central, mock_client = await central_local_factory.get_central(TEST_DEVICES)
+    central, mock_client = await central_local_factory.get_default_central(TEST_DEVICES)
     switch: HmSwitch = cast(
         HmSwitch, await get_generic_entity(central, "VCU2128127:4", "STATE")
     )
@@ -154,7 +154,7 @@ async def test_generic_wrapped_entity(
     central_local_factory: helper.CentralUnitLocalFactory,
 ) -> None:
     """Test wrapped entity."""
-    central, mock_client = await central_local_factory.get_central(TEST_DEVICES)
+    central, mock_client = await central_local_factory.get_default_central(TEST_DEVICES)
     wrapped_entity: HmSensor = cast(
         HmSensor, await get_wrapper_entity(central, "VCU3609622:1", "LEVEL")
     )