+++ tests/test_lock.py
@@ -23,7 +23,7 @@ async def test_cerflock(
     central_local_factory: helper.CentralUnitLocalFactory,
 ) -> None:
     """Test CeRfLock."""
-    central, mock_client = await central_local_factory.get_central(TEST_DEVICES)
+    central, mock_client = await central_local_factory.get_default_central(TEST_DEVICES)
     lock: CeRfLock = cast(CeRfLock, await get_custom_entity(central, "VCU0000146", 1))
     assert lock.usage == HmEntityUsage.CE_PRIMARY
 
@@ -74,7 +74,7 @@ async def test_ceiplock(
     central_local_factory: helper.CentralUnitLocalFactory,
 ) -> None:
     """Test CeIpLock."""
-    central, mock_client = await central_local_factory.get_central(TEST_DEVICES)
+    central, mock_client = await central_local_factory.get_default_central(TEST_DEVICES)
     lock: CeIpLock = cast(CeIpLock, await get_custom_entity(central, "VCU9724704", 1))
     assert lock.usage == HmEntityUsage.CE_PRIMARY