+++ tests/test_switch.py
@@ -23,7 +23,7 @@ async def test_ceswitch(
     central_local_factory: helper.CentralUnitLocalFactory,
 ) -> None:
     """Test CeSwitch."""
-    central, mock_client = await central_local_factory.get_central(TEST_DEVICES)
+    central, mock_client = await central_local_factory.get_default_central(TEST_DEVICES)
     switch: CeSwitch = cast(CeSwitch, await get_custom_entity(central, "VCU2128127", 4))
     assert switch.usage == HmEntityUsage.CE_PRIMARY
 
@@ -73,7 +73,7 @@ async def test_hmswitch(
     central_local_factory: helper.CentralUnitLocalFactory,
 ) -> None:
     """Test HmSwitch."""
-    central, mock_client = await central_local_factory.get_central(TEST_DEVICES)
+    central, mock_client = await central_local_factory.get_default_central(TEST_DEVICES)
     switch: HmSwitch = cast(
         HmSwitch, await get_generic_entity(central, "VCU2128127:4", "STATE")
     )
@@ -124,7 +124,9 @@ async def test_hmsysvarswitch(
     central_local_factory: helper.CentralUnitLocalFactory,
 ) -> None:
     """Test HmSysvarSwitch."""
-    central, mock_client = await central_local_factory.get_central({}, add_sysvars=True)
+    central, mock_client = await central_local_factory.get_default_central(
+        {}, add_sysvars=True
+    )
     switch: HmSysvarSwitch = cast(
         HmSysvarSwitch, await get_sysvar_entity(central, "sv_alarm_ext")
     )