+++ tests/test_text.py
@@ -20,7 +20,7 @@ async def no_test_hmtext(
     central_local_factory: helper.CentralUnitLocalFactory,
 ) -> None:
     """Test HmText. There are currently no text entities"""
-    central, mock_client = await central_local_factory.get_central(TEST_DEVICES)
+    central, mock_client = await central_local_factory.get_default_central(TEST_DEVICES)
     text: HmText = cast(
         HmText, await get_generic_entity(central, "VCU7981740:1", "STATE")
     )
@@ -32,7 +32,9 @@ async def test_hmsysvartext(
     central_local_factory: helper.CentralUnitLocalFactory,
 ) -> None:
     """Test HmSysvarText. There are currently no text entities"""
-    central, mock_client = await central_local_factory.get_central({}, add_sysvars=True)
+    central, mock_client = await central_local_factory.get_default_central(
+        {}, add_sysvars=True
+    )
     text: HmSysvarText = cast(
         HmSysvarText, await get_sysvar_entity(central, "sv_string_ext")
     )